Bug 1255158 - Update graphite2 library to release 1.3.7. r=jdaggett, a=lizzard
--HG--
extra : source : e9558895348ab442a0726f84f17d45ce6471c7d2

+++ gfx/graphite2/src/inc/GlyphFace.h
@@ -51,7 +51,7 @@ public:
     const Position    & theAdvance() const;
     const Rect        & theBBox() const { return m_bbox; }
     const sparse      & attrs() const { return m_attrs; }
-    uint16              getMetric(uint8 metric) const;
+    int32               getMetric(uint8 metric) const;
 
     CLASS_NEW_DELETE;
 private:
